plant giving me allergies

+++ src/app/services/product.service.ts
@@ -1,4 +1,4 @@
-import { DialogService } from '@services/shared/dialog/dialog.service';
+import { NotificationsService } from './shared/notifications/notifications.service';
 import { AngularFireStorage } from '@angular/fire/storage';
 import { Injectable } from '@angular/core';
 import { AngularFirestore } from '@angular/fire/firestore';
@@ -12,7 +12,7 @@ export class ProductService {
   constructor(
     private firestore: AngularFirestore,
     private storage: AngularFireStorage,
-    private dialogService: DialogService
+    private notificationsService: NotificationsService
   ) {}
 
   getCollection(collection) {
@@ -31,36 +31,45 @@ export class ProductService {
       );
   }
 
-  getProduct(id, collection) {
+  getProduct(id: string, collection: string) {
     return this.firestore.doc(`${collection}/${id}`);
   }
 
-  createProduct(product: ProductModel, collection) {
-    return this.firestore.collection(collection).add(product);
+  createProduct(product: ProductModel, collection: string): void {
+    this.firestore
+      .collection(collection)
+      .add(product)
+      .catch((err) => console.log('error creating product', err));
+    this.notificationsService.successAlert(
+      `${product.name} successfully created`
+    );
   }
 
-  updateProduct(product: ProductModel, collection) {
-    this.firestore.doc(`${collection}/${product.id}`).update(product);
+  updateProduct(product: ProductModel, collection: string): void {
+    this.firestore
+      .doc(`${collection}/${product.id}`)
+      .update(product)
+      .catch((err) => console.log('error updating product', err));
+    this.notificationsService.successAlert(
+      `${product.name} successfully updated`
+    );
   }
 
-  deleteProduct(product: ProductModel, collection) {
-    console.log(`deleting ${collection}/${product.id}`);
-    this.dialogService
-      .openDialog({ action: 'Delete' })
-      .afterClosed()
-      .subscribe((res) => {
-        if (res.event !== 'Cancel') {
-          this.firestore
-            .doc(`${collection}/${product.id}`)
-            .delete()
-            .catch((err) => console.log('error deleting that product'));
-          this.removeStorageRef(product.imageUrl);
-        }
-      });
+  deleteProduct(product: ProductModel, collection: string): void {
+    this.firestore
+      .doc(`${collection}/${product.id}`)
+      .delete()
+      .catch((err) => console.log('error deleting that product', err));
+    this.removeStorageRef(product.imageUrl);
+    this.notificationsService.successAlert(
+      `${product.name} successfully deleted`
+    );
   }
 
-  removeStorageRef(imgUrl): void {
-    const ref = this.storage.refFromURL(imgUrl);
-    ref.delete();
+  removeStorageRef(imgUrl: string): void {
+    if (imgUrl) {
+      const ref = this.storage.refFromURL(imgUrl);
+      ref.delete();
+    }
   }
 }
